requirements.txt
numpy==1.21.3
pandas==1.3.5
Pillow==8.4.0
progressbar==2.5
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
torch==1.9.0
torchvision==0.9.0
typing-extensions==3.10.0.2
